updated client_id in the database (just to be larger), pulled some dead code, forced IE compatability view, and fixed password reset. I had done this all before but apparently my changes were being pushed to the wrong branch, and I lost them.)

+++ echelon.sql
@@ -89,7 +89,7 @@ CREATE TABLE `ech_logs` (
   `id` smallint(5) unsigned NOT NULL AUTO_INCREMENT,
   `type` varchar(64) DEFAULT NULL,
   `msg` varchar(255) DEFAULT '',
-  `client_id` smallint(5) DEFAULT NULL,
+  `client_id` int(32) DEFAULT NULL,
   `user_id` smallint(5) DEFAULT NULL,
   `time_add` int(32) DEFAULT NULL,
   `game_id` mediumint(10) DEFAULT NULL,
